requirements.txt
absl-py==2.3.1
attrs==25.3.0
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
scipy==1.16.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3